add class ForceSensor to manage force value; data flow: force_core-->servo_core-->controller-->pc.(RPC channel and Publish channel have passed); next to debug the force calibration and compensation

+++ Motion_controller/src/controller/src/update_element.cpp
@@ -104,7 +104,7 @@ void ControllerPublish::updateTorqueFdb()
 {
 	torque_fdb_.data_count = 6;
 	
-	force_sensor_ptr_->updateSourceValue(GROUP_0, &torque_fdb_.data[0], torque_fdb_.data_count);
+	force_sensor_ptr_->updateSourceValue(GROUP_0, &torque_fdb_.data[0], (int)torque_fdb_.data_count);
 	
 	//force_sensor_ptr_->calibratedForceSensor(GROUP_0);
